+++ changelog.md
@@ -5,6 +5,9 @@
 - Fixed rare null deref caused by mastery ritual code querying GameManager without a valid GameManager instance
 - Added mastery descriptions to [ItemTips](https://enter-the-gungeon.thunderstore.io/package/Glorfindel/ItemTips/?utm_source=discord) output
 - Fixed issue with Alyx completely decaying immediately after spawning in under unknown circumstances
+- Reworked Hallaeribut's Famished state
+	- Old behavior: feeds on the player every 30 seconds while Famished, no ammo gain
+	- New behavior: feeds on the player when reduced to 0 ammo and refills 20% ammo immediately
 
 ## 1.30.0 (2026-05-07)